Try to implement linked list with C. Refactoring tty and console. Add input buffer for keyboard.

+++ src/include/libc/string/string.c
@@ -33,7 +33,7 @@ char *strcpy(char *dest, const char *src){
 }
 
 void backspace(char *src){
-		int32_t len = strlen(src);
+		size_t len = strlen(src);
 		src[len-1] = '\0';
 }
